app/requirements.txt
Django==5.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.10
drf_yasg==1.21.7
djangorestframework==3.15.1
Pillow==9.5.0
djangorestframework-simplejwt==5.3.0
hiredis==3.0.0
redis==5.1.1